Fixed bugs in getdirn and added getdirn2
Getdirn code had a few bugs which were fixed
1. Order of multiplication of rotation matrices was inverted - for Rb2c and Ri2carr
2. Ri2b was unnecessarily being flipped for yaw and pitch because of order
multiplication switch. Fixed that and removed this.
3. plxpc1 2 3 were wrongly computed in camera coordinates (as sign of the
angles was not being accounted for). Now a separate function getvect1t2 added
to do this.
4. thetvra1 2 3 wrongly computed. By definition vzs to ra has inverted angles.
Corrected this.

Added getdirn2.py to compute ratios for all sources in srclist. getdirn.py
is now just a sample code for the way computation is performed. The lightcurve
fitting method is abandoned. Instead flux ratios (in two LAXPCs) are being used (computed in
getdirn2.py). The plan is to have another getdirn3.py which can obtain location
regions for given ratios.